Move code to openfl.federated.data.sources
