+++ openfl/federated/data/sources/__init__.py
@@ -1,0 +1,10 @@
+# Copyright 2020-2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+"""This package contains utilities for verifying datasets."""
+
+from openfl.federated.data.sources.data_source import DataSource, DataSourceType
+from openfl.federated.data.sources.local_data_source import LocalDataSource
+from openfl.federated.data.sources.verifiable_dataset_info import (
+    DatasetFormat,
+    VerifiableDatasetInfo,
+)
