+++ openfl/federated/data/sources/pytorch/__init__.py
@@ -3,9 +3,9 @@
 
 """This package contains utilities for PyTorch verifying datasets."""
 
-from openfl.utilities.verifdataset.pytorch.verifiable_map_style_dataset import (
+from openfl.federated.data.sources.pytorch.verifiable_map_style_dataset import (
     VerifiableMapStyleDataset,
 )
-from openfl.utilities.verifdataset.pytorch.verifiable_map_style_image_folder import (
+from openfl.federated.data.sources.pytorch.verifiable_map_style_image_folder import (
     VerifiableImageFolder,
 )
