+++ openfl/federated/data/sources/pytorch/local_folder.py
