+++ openfl/federated/data/sources/pytorch/local_image_folder.py
@@ -6,7 +6,9 @@ from typing import Union
 
 from PIL import Image
 
-from openfl.utilities.verifdataset.pytorch.local_folder import LocalFolder
+from openfl.federated.data.sources.pytorch.local_folder import LocalFolder
+
+# from openfl.federated.data.sources.pytorch.local_folder import LocalFolder
 
 
 class LocalImageFolder(LocalFolder):
