+++ openfl/federated/data/sources/verifiable_dataset_info.py
@@ -9,8 +9,8 @@ from hashlib import sha384
 from pathlib import Path
 from typing import List
 
-from openfl.utilities.verifdataset.data_source import DataSource, DataSourceType
-from openfl.utilities.verifdataset.local_data_source import LocalDataSource
+from openfl.federated.data.sources.data_source import DataSource, DataSourceType
+from openfl.federated.data.sources.local_data_source import LocalDataSource
 
 
 class DatasetFormat(Enum):
