--- openfl/utilities/verifdataset/__init__.py
@@ -1,10 +1,0 @@
-# Copyright 2020-2025 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-"""This package contains utilities for verifying datasets."""
-
-from openfl.utilities.verifdataset.data_source import DataSource, DataSourceType
-from openfl.utilities.verifdataset.local_data_source import LocalDataSource
-from openfl.utilities.verifdataset.verifiable_dataset_info import (
-    DatasetFormat,
-    VerifiableDatasetInfo,
-)
